fix: 이벤트 요청 위치변경, id계산 함수 수정
