+++ src/app/(auth)/_components/calendar/index.tsx
@@ -1,12 +1,7 @@
 "use client";
 
-import { createClient } from "@/libs/supabase/client";
 import useCalendarState from "@/store/calendarDay";
-import useOptionState from "@/store/options";
-import { getOptionIdofPath } from "@/utils/path";
 import Image from "next/image";
-import { usePathname } from "next/navigation";
-import { useEffect } from "react";
 import tw from "tailwind-styled-components";
 import ContentsBox from "../contentsBox";
 
@@ -39,35 +34,6 @@ export default function Calendar({ children }: Prop) {
   const weekdays = ["일", "월", "화", "수", "목", "금", "토"];
   const { viewDate, goToLeftMonth, goToRightMonth, goToTodayMonth } =
     useCalendarState();
-
-  const path = usePathname();
-  const { options } = useOptionState();
-
-  const supabase = createClient();
-
-  useEffect(() => {
-    const optionId = getOptionIdofPath(path, options);
-
-    if (!optionId) {
-      console.error("주소가 유효하지 않습니다.");
-      return;
-    }
-
-    const getEvents = async () => {
-      const { data, error } = await supabase
-        .from("events")
-        .select()
-        .eq("option_id", optionId);
-
-      if (error) {
-        console.log("조회오류", error);
-      } else {
-        console.log("성공", data);
-      }
-    };
-
-    getEvents();
-  }, [path]);
 
   return (
     <ContentsBox>
